refactor: User의 Position 원시값 포장
